Reorganize package structure.
